backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy
databases
Pillow
openai
